fix label

+++ src/boltz/model/model.py
@@ -292,7 +292,7 @@ class Boltz1(LightningModule):
             multiplicity_diffusion_train=-1,
             diffusion_samples=-1,
         )
-        label = torch.tensor(batch["label"]).unsqueeze(dim=0)
+        label = torch.tensor(batch["label"]).unsqueeze(dim=0).to(torch.bfloat16)
         print(out, label)
         loss = F.binary_cross_entropy(out, label)
         self.log("train/loss", loss)
